Rename _memory.py into _linux_memory.py
